requirements.txt
pyhcl==0.4.5